+++ backend/html_report/ui/src/components/distilled/metric/metric.ts
@@ -1,6 +1,7 @@
 import { customElement, property } from 'lit/decorators.js'
 import { Statistics, dummyStatistics } from '../../../model/statistics';
 import { LitElement, html } from 'lit'
+import { CreateMetricOptions, Metric, dummyMetric } from '../../../model/metric';
 
 
 
@@ -13,13 +14,13 @@ statistics: Statistics = dummyStatistics;
 openView = false;
 
 @property()
-label: string = "";
+position: number = 0;
 
 @property()
-value: number = 0;
+metric: Metric = dummyMetric
 
 @property()
-explanation: string = "";
+chosenMethod = CreateMetricOptions.SetManually
     
   constructor() {
     super()
@@ -34,20 +35,44 @@ explanation: string = "";
     this.openView = !this.openView;
   }
 
+  addMetric() {
+
+    this.dispatchEvent(new CustomEvent("UseMetric", {
+      composed: true,
+      bubbles: true,
+      detail: {
+          metric: this.metric
+      }
+    }))
+  }
+  
+  editMetric() {
+    this.dispatchEvent(new CustomEvent("EditMetric", {
+      composed: true,
+      bubbles: true,
+      detail: {
+          position: this.position,
+          metric: this.metric
+      }
+    }))
+  }
+
   render() {
     // <p>${this.Rating ? "Good" : "Bad"}</p>
     return html`
         <tr class="row second">
         <td class="column">
-        <h3>${this.label}</h3>
+        <h3>${this.metric.label}</h3>
         </td>
         <td class="column">
-        <p>${this.value}</p>
+        <p>${this.metric.value}</p>
         </td>
         <td class="column">
         none
         </td>
         <td class="column">
+        ${this.chosenMethod === CreateMetricOptions.FromOthers ?
+          html`<button class="include-metric" @click=${this.addMetric}>+</button>` : ""}
         <button
         class="more ${this.openView ? "turn" : ""}"
         @click=${this.handleClick}
@@ -57,7 +82,8 @@ explanation: string = "";
         
         <tr class="row info ${this.openView ? "show" : ""}">
         <td >
-        <p>${this.explanation}</p>
+        <p>${this.metric.explanation}</p>
+        <button class="edit-metric" @click=${this.editMetric}>Edit</button>
         </td>
         </tr>
     `