+++ backend/html_report/ui/src/components/doc/doc.css.ts
@@ -18,7 +18,7 @@ export default css`
     transform: translateX(50%);
     top: 5%;
     background-color: var(--gray30);
-    padding: 1.2rem 3.2rem;
+    padding: var(--spacingHalf) var(--spacing);
     font-size: 2rem;
     border-radius: 10px;
     color: var(--80);